AdamTheAnalyst: Added multiple profiles, config file, shell break.
I have added the following, this should be enough to get me up and building profiles:

	- Config file
Found in yuuki.conf, allows users to define the location of profiles and network settings without passing them as args. This config then gets loaded into the flask app context so can be accessed by profiles if necassary in the future (IP's of routers etc) using "from flask import current_app as app". I have stored this in the config dict ["yuuki"] rather than using flask's traditional "app.config.from_envvar" to load it into Flask's default config. This means that the user can't define profile names that would interfer with Flask's operation. If you don't like this get me to change it.

	- Break in shell.py
Just to catch ^C's, it wasn't exiting before

	- Multiple profile loading
Profiles are now defined in the config file under the "[profiles]" section, they are then loaded in to a list which gets passed to the Dispatcher's constructor and loaded in the same way.
 master