AdamTheAnalyst (Fixes): Added Optional Switches, Removed Break, Added Warning:
Changes:

Added the warning about overlapping profile actions

Removed my break on ^C

I have added back in the ability to specify profiles from the command line, this by default binds to 127.0.0.1 as it will be used for dev/testing.

If a valid config fil is specified and a port switch is passed, the config file takes priority.